+++ src/components/VoteContainer.tsx
